Safari-611.1.10/LayoutTests/fast/innerHTML/004.xhtml

<?xml version="1.0" encoding="UTF-8"?>
<!DOCTYPE html PUBLIC
    "-//W3C//DTD XHTML 1.1 plus MathML 2.0 plus SVG 1.1//EN"
    "http://www.w3.org/2002/04/xhtml-math-svg/xhtml-math-svg-flat.dtd">
<html xmlns="http://www.w3.org/1999/xhtml">
<head>
<meta name="description" content="This tests singular elements too" />
<title>xhtml innerHTML test</title>
</head>
<body>
<span>Content:</span>
<pre id="content">placeholder</pre>
<div></div>
<span></span>
<b>test</b>
<br />
<div></div>
<svg xmlns="http://www.w3.org/2000/svg" xmlns:xlink="http://www.w3.org/1999/xlink" width="300" height="200">
  <?xml-stylesheet type="text/xsl" href=""?>
  <circle cx="150" cy="100" r="50" xlink:title="test"/>
</svg>
<script>
window.addEventListener ("load", function() {
    if (window.testRunner)
	    testRunner.dumpAsText();
    document.getElementById("content").firstChild.nodeValue = document.documentElement.innerHTML;
}, false);

</script>
</body>
</html>
