LayoutTests/fast/innerHTML/004-expected.txt

Content:

<head>
<meta name="description" content="This tests singular elements too" />
<title>xhtml innerHTML test</title>
</head>
<body>
<span>Content:</span>
<pre id="content">placeholder</pre>
<div></div>
<span></span>
<b>test</b>
<br />
<div></div>
<svg xmlns="http://www.w3.org/2000/svg" xmlns:xlink="http://www.w3.org/1999/xlink" width="300" height="200">
  <?xml-stylesheet type="text/xsl" href=""?>
  <circle cx="150" cy="100" r="50" xlink:title="test"/>
</svg>
<script>
if (window.layoutTestController)
	layoutTestController.dumpAsText();
document.getElementById("content").firstChild.nodeValue = document.documentElement.innerHTML;
</script>
</body>
test 
